# Speedup Rust Execution