--- a/rust/src/main.rs
+++ b/rust/src/main.rs
@@ -3,15 +3,17 @@ mod levenshtein_distance;
 use levenshtein_distance::levenshtein_distance;
 
 fn main() {
-    let lines: Vec<&str> = include_str!("../../sample.txt").split('\n').collect();
+
+    let mut lines: Vec<&str> = include_str!("../../sample.txt").split('\n').collect();
+    lines.insert(0, "");
+    let lines: Vec<Vec<char>> = lines.iter().map(|line| line.chars().collect()).collect();
 
     let benchmark = || {
         for _ in 0..10000 {
-            let mut last_value = "";
-            for line in &lines {
-                levenshtein_distance(last_value, line);
-                last_value = line;
-            }
+            lines.iter().zip(lines[1..].iter())
+                .for_each(|(a, b)| {
+                    levenshtein_distance(a, b);
+                });
         }
     };
 
@@ -27,8 +29,8 @@ fn main() {
     print!("{}", sec);
 
     // check
-    let answers: Vec<String> = (0..lines.len()-1)
-        .map(|i| levenshtein_distance(lines[i], lines[i+1]))
+    let answers: Vec<String> = (1..lines.len()-1)
+        .map(|i| levenshtein_distance(&lines[i], &lines[i+1]))
         .map(|dist| dist.to_string())
         .collect();
     eprintln!("{}", answers.join(","));